cephadm: move haproxy and keepalived classes to a new file
