+++ src/cephadm/cephadmlib/daemons/__init__.py
@@ -1,4 +1,5 @@
 from .custom import CustomContainer
 from .tracing import Tracing
+from .ingress import HAproxy, Keepalived
 
-__all__ = ['CustomContainer', 'Tracing']
+__all__ = ['CustomContainer', 'Tracing', 'HAproxy', 'Keepalived']
